+++ Lab1/LineSegment.h
@@ -4,14 +4,14 @@ class LineSegment
 {public:
 	Point2D A, B;
 
-	void SetA(float x, float y);
-	void SetB(float x, float y);
+	void SetA(int x, int y);
+	void SetB(int x, int y);
 
 	Point2D GetA();
 	Point2D GetB();
 
 	LineSegment();
-	LineSegment(float xa, float ya, float xb, float yb);
+	LineSegment(int xa, int ya, int xb, int yb);
 
 	void DrawLine(sf::RenderWindow *target);
 };
