+++ Lab1/Point2D.h
@@ -5,15 +5,15 @@ class Point2D
 public:
 	float x, y;
 
-	void SetX(float x);
-	void SetY(float y);
+	void SetX(int x);
+	void SetY(int y);
 
 	float GetX();
 	float GetY();
 
 
 	Point2D();
-	Point2D(float x, float y);
+	Point2D(int x, int y);
 	void DrawPoint(sf::RenderWindow *target);
 };
